fix: fix bug for category detail page, forgot password and display dine-in order detail
- Fix bug for category detail page that cannot update without upload image
- Fix verification reset password token
- Fix cannot complete order or reservation that has canceled item

+++ backend/src/main/java/com/phucx/phucxfandb/service/category/impl/CategoryUpdateServiceImpl.java
@@ -39,7 +39,7 @@ public class CategoryUpdateServiceImpl implements CategoryUpdateService {
         // upload new image
         if(requestCategoryDTO.getPicture()!=null && !requestCategoryDTO.getPicture().isEmpty()){
             String newImageName = ImageUtils.extractImageNameFromUrl(requestCategoryDTO.getPicture());
-            if(existingCategory.getPicture() != null ){
+            if(existingCategory.getPicture() != null && !existingCategory.getPicture().isBlank() ){
                 if(!newImageName.equalsIgnoreCase(existingCategory.getPicture())){
                     imageUpdateService.removeImages(List.of(existingCategory.getPicture()));
                     requestCategoryDTO.setPicture(newImageName);
